fix(#2, #8): Use idea cache folder to prevent permission problems
In the past there were several issues stating that the plugins folder provided by the IntelliJ SDK is not writable.
To avoid this for the future and specific setup the plugins path (PathManager#getPluginsPath) is used instead of PluginPathManager#getPluginHome, as stated in https://intellij-support.jetbrains.com/hc/en-us/articles/206544519-Directories-used-by-the-IDE-to-store-settings-caches-plugins-and-logs, the system folder should always be writable by the plugin.